+++ cfgov/v1/models/__init__.py
@@ -4,6 +4,7 @@ from .events import *
 from .molecules import *
 from .organisms import *
 from .snippets import *
+from .landing_page import *
 
 if settings.DEBUG:
     from .demo import *