+++ cfgov/v1/models/molecules.py
@@ -5,6 +5,8 @@ from wagtail.wagtailimages.blocks import ImageChooserBlock
 
 from . import atoms
 
+from ..util import id_validator
+
 
 def isRequired(field_name):
     return [str(field_name) + ' is required.']
@@ -12,121 +14,112 @@ def isRequired(field_name):
 
 class HalfWidthLinkBlob(blocks.StructBlock):
     heading = blocks.CharBlock(max_length=100, required=True)
-    content = blocks.RichTextBlock(blank=True)
-    links = blocks.ListBlock(blocks.StructBlock([
-        ('text', blocks.CharBlock(required=False)),
-        ('url', blocks.URLBlock(required=False)),
-    ], icon='user', required=False)
-    )
+    body = blocks.RichTextBlock(blank=True)
+    links = blocks.ListBlock(atoms.Hyperlink(), required=False)
 
     class Meta:
         icon = 'link'
-        template = 'v1/demo/molecules/half_width_link_blob.html'
+        template = '_includes/molecules/half-width-link-blob.html'
 
 
 class ImageText5050(blocks.StructBlock):
-    title = blocks.CharBlock(max_length=100, required=True)
-    description = blocks.RichTextBlock(blank=True)
-    image = ImageChooserBlock(required=False)
-    image_path = blocks.CharBlock(required=False)
-    image_alt = blocks.CharBlock(required=False)
-    is_widescreen = blocks.BooleanBlock(required=False)
-    is_button = blocks.BooleanBlock(required=False)
-    link_url = blocks.URLBlock(required=False)
-    link_text = blocks.CharBlock(max_length=100, required=False)
-
-    def clean(self, data):
-        error_dict = {}
-        try:
-            block_data = super(ImageText5050, self).clean(data)
-        except ValidationError as e:
-            error_dict.update(e.params)
-            block_data = data
-
-        if not block_data['image'] and not block_data['image_path'] and not block_data['image_alt']:
-            img_err = ['Please upload or enter an image path']
-            error_dict.update({'image': img_err, 'image_path': img_err, 'image_alt': isRequired('Image alt')})
-
-        if block_data['image'] and block_data['image_path']:
-            img_err = ['Please select one method of image rendering']
-            error_dict.update({
-                'image': img_err,
-                'image_path': img_err})
-
-        if block_data['image_path'] and not block_data['image_alt']:
-            error_dict.update({'image_alt': isRequired('Image Alt')})
-
-        if error_dict:
-            raise ValidationError("ImageText5050 validation errors", params=error_dict)
-        else:
-            return block_data
+    heading = blocks.CharBlock(max_length=100, required=True)
+    body = blocks.RichTextBlock(blank=True)
+    image = atoms.ImageBasic()
+    is_widescreen = blocks.BooleanBlock(required=False, label="Use 16:9 image")
+    is_button = blocks.BooleanBlock(required=False, label="Show links as button")
+    links = blocks.ListBlock(atoms.Hyperlink(), required=False)
 
     class Meta:
         icon = 'image'
-        template = 'v1/demo/molecules/image_text_5050.html'
+        template = '_includes/molecules/image-text-50-50.html'
 
 
 class ImageText2575(blocks.StructBlock):
     heading = blocks.CharBlock(max_length=100, required=True)
     body = blocks.RichTextBlock(required=True)
-    image = ImageChooserBlock(required=False)
-    image_path = blocks.CharBlock(required=False)
-    image_alt = blocks.CharBlock(required=False)
-    link_url = blocks.URLBlock(required=False)
-    link_text = blocks.CharBlock(max_length=100, required=False)
+    image = atoms.ImageBasicAlt()
+    links = blocks.ListBlock(atoms.Hyperlink(), required=False)
     has_rule = blocks.BooleanBlock(required=False)
-
-    def clean(self, data):
-        error_dict = {}
-        try:
-            block_data = super(ImageText5050, self).clean(data)
-        except ValidationError as e:
-            error_dict.update(e.params)
-            block_data = data
-
-        if not block_data['image'] and not block_data['image_path'] and not block_data['image_alt']:
-            img_err = ['Please upload or enter an image path']
-            error_dict.update({'image': img_err, 'image_path': img_err, 'image_alt': isRequired('Image alt')})
-
-        if block_data['image'] and block_data['image_path']:
-            img_err = ['Please select one method of image rendering']
-            error_dict.update({
-                'image': img_err,
-                'image_path': img_err})
-
-        if block_data['image_path'] and not block_data['image_alt']:
-            error_dict.update({'image_alt': isRequired('Image Alt')})
-
-        if error_dict:
-            raise ValidationError("ImageText2575 validation errors", params=error_dict)
-        else:
-            return block_data
 
     class Meta:
         icon = 'image'
-        template = 'v1/wagtail/molecules/image_text_2575.html'
+        template = '_includes/molecules/image-text-25-75.html'
 
 
 class TextIntroduction(blocks.StructBlock):
     heading = blocks.CharBlock(max_length=100, required=True)
-    intro = blocks.CharBlock(max_length=100, required=True)
+    intro = blocks.CharBlock(max_length=100, required=False)
     body = blocks.RichTextBlock(required=False)
-    link_url = blocks.URLBlock(required=False)
-    link_text = blocks.CharBlock(max_length=100, required=False)
+    link = atoms.Hyperlink(required=False)
     has_rule = blocks.BooleanBlock(required=False)
 
     class Meta:
         icon = 'title'
-        template = 'v1/demo/molecules/text_introduction.html'
+        template = '_includes/molecules/text-introduction.html'
+
+
+class Hero(blocks.StructBlock):
+    heading = blocks.CharBlock(max_length=100, required=True)
+    body = blocks.RichTextBlock(required=False)
+
+    image = atoms.ImageBasic()
+
+    background_color = blocks.CharBlock(max_length=100, required=False)
+    link = atoms.Hyperlink()
+    is_button = blocks.BooleanBlock(required=False)
+
+    class Meta:
+        icon = 'image'
+        template = '_includes/molecules/hero.html'
+
+
+class FormFieldWithButton(blocks.StructBlock):
+    btn_text = blocks.CharBlock(max_length=100, required=True)
+
+    required = blocks.BooleanBlock(required=False)
+    id = blocks.CharBlock(max_length=100, required=True)
+    info = blocks.RichTextBlock(required=False, label="Disclaimer")
+    label = blocks.CharBlock(max_length=100, required=True)
+    type = blocks.ChoiceBlock(choices=[
+        ('text', 'Text'),
+        ('checkbox', 'Checkbox'),
+        ('email', 'Email'),
+        ('number', 'Number'),
+        ('url', 'URL'),
+        ('radio', 'Radio'),
+    ], icon='cup', required=True)
+    placeholder = blocks.CharBlock(max_length=100, required=False)
+
+    def clean(self, data):
+        error_dict = {}
+
+        try:
+            data = super(FormFieldWithButton, self).clean(data)
+        except ValidationError as e:
+            error_dict.update(e.params)
+
+        if not id_validator(data['id']):
+            id_err = ['Id must only contain alphabets, numbers, underscores and hyphens']
+            error_dict.update({'id': id_err})
+
+        if error_dict:
+            raise ValidationError("ImageBasicUrlAlt validation errors", params=error_dict)
+        else:
+            return data
+
+    class Meta:
+        icon = 'mail'
+        template = '_includes/molecules/form-field-with-button.html'
 
 
 class CallToAction(blocks.StructBlock):
-    slug = blocks.CharBlock(required=True)
-    paragraph = blocks.RichTextBlock()
+    slug_text = blocks.CharBlock(max_length=100, required=True)
+    paragraph_text = blocks.RichTextBlock(required=True)
     button = atoms.Hyperlink()
 
     class Meta:
-        template = 'v1/wagtail/molecules/call-to-action.html'
+        template = '_includes/molecules/call-to-action.html'
         icon = 'grip'
         label = 'Call to Action'
 
@@ -140,17 +133,17 @@ class ContactAddress(blocks.StructBlock):
     zip_code = blocks.CharBlock(max_length=15, required=False)
 
     class Meta:
-        template = 'v1/wagtail/molecules/contact-address.html'
+        template = '_includes/molecules/contact-address.html'
         icon = 'mail'
         label = 'Address'
 
 
 class ContactEmail(blocks.StructBlock):
-    emails = blocks.ListBlock(atoms.Hyperlink(label='Email'))
+    emails = blocks.ListBlock(atoms.Hyperlink())
 
     class Meta:
         icon = 'mail'
-        template = 'v1/wagtail/molecules/contact-email.html'
+        template = '_includes/molecules/contact-email.html'
         label = 'Email'
 
 
@@ -166,5 +159,5 @@ class ContactPhone(blocks.StructBlock):
 
     class Meta:
         icon = 'mail'
-        template = 'v1/wagtail/molecules/contact-phone.html'
+        template = '_includes/molecules/contact-phone.html'
         label = 'Phone'