+++ cfgov/v1/models/organisms.py
@@ -1,9 +1,31 @@
 from wagtail.wagtailcore import blocks
+from wagtail.wagtailcore.fields import RichTextField, StreamField
+from wagtail.wagtailadmin.edit_handlers import FieldPanel, StreamFieldPanel
+
+from . import molecules
 
 
 class Well(blocks.StructBlock):
-    template_path = blocks.CharBlock(required=True)
+    content = blocks.RichTextBlock(required=True)
 
     class Meta:
-        icon = 'link'
-        template = 'v1/demo/organisms/well.html'
+        icon = 'title'
+        template = '_includes/organisms/well.html'
+
+
+class EmailSignUp(blocks.StructBlock):
+    heading = blocks.CharBlock(max_length=100, required=True)
+    text = blocks.CharBlock(required=True)
+    gd_code = blocks.CharBlock(required=False)
+
+    form_field = blocks.ListBlock(molecules.FormFieldWithButton(), icon='mail', required=False)
+
+    panels = [
+        FieldPanel('text'),
+        FieldPanel('gd_code'),
+        FieldPanel('form_field'),
+    ]
+
+    class Meta:
+        icon = 'mail'
+        template = '_includes/organisms/email-signup.html'