+++ CHANGELOG.md
@@ -94,6 +94,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Added WP Import Data Contact processor
 - Added templates and CSS for the Adding Sidebar Breakout organism.
 - Added cf-tables and tables molecule
+- Landing Page Type
 
 ### Changed
 - Updated the primary nav to move focus as user enters and leaves nav levels
@@ -155,6 +156,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Changes .env Project configuration workon control flow to direct stdout and stderr to /dev/null.
 - Upgrade wagtail to 1.2
 - Cleaned up and rebuilt the secondary nav to reduce complexity and fix bugs
+- Routed landing page type related molecules and organisms to use jinja2/v1/_includes/ template locations
 
 ### Removed
 - Removed unused exportsOverride section,